requirements.txt
aerich==0.5.4
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiosqlite==0.16.1
appdirs==1.4.4
asgiref==3.4.1
astroid==2.5.6
async-timeout==3.0.1
asyncpg==0.23.0
asynctest==0.13.0
attrs==21.2.0
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
coverage==5.5
cryptography==3.4.7
ddlparse==1.10.0
dictdiffer==0.8.1
ecdsa==0.17.0
fastapi==0.66.0
gevent==21.8.0
greenlet==1.1.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1
iso8601==0.1.16
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pylint==2.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
pytest-asyncio==0.15.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==1.1.0
tortoise-orm==0.17.5
types-aiofiles==0.1.7
types-orjson==3.6.0
types-pytz==2021.1.2
types-PyYAML==5.4.6
types-ujson==0.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
websocket-client==1.2.1
websockets==9.1
wrapt==1.12.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
